requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
Babel==2.14.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==4.2.6
django-minio-storage==0.5.7
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
fonttools==4.49.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
Markdown==3.5.2
minio==7.2.0
packaging==23.2
pdfkit==1.0.0
phonenumbers==8.13.26
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.19.0
pydyf==0.8.0
pyphen==0.14.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
weasyprint==61.0
webencodings==0.5.1
zopfli==0.2.3
